Add an override for weka-dev
We don't want to update pom-scijava to 3.9.1, since it causes
compatibility issues with TWS. Don't tell us to do anything to
weka-dev until/unless more changes happen to it.

This feels a little hacky, since it's not really changes to weka-dev
SNAPSHOT artifacts to which we want to react, but rather new weka-dev
_releases_. However, I didn't implement a release version override
yet. So we'll piggyback on the "last vetted" feature as is for now.