newest-releases.sh: use a rules config file
This configuration allows to exclude certain version patterns from
consideration when analyzing versions of components. For example,
if the latest version of a third party component is an alpha release, we
probably do not want to report that component as needing an upgrade.

This slickness allows us to stop manually and repeatedly vetting third
party components, in favor of rules for when upgrades may be warranted.

See: https://www.mojohaus.org/versions-maven-plugin/version-rules.html