html-report: fix bug with erroneous None action
I wanted to make it so that timestamp overrides automatically set the
action to None, so that maintainers could decide to postpone updating
to newly released versions. This is useful when major a new release
lands for a third-party component, since there is often not time
immediately to upgrade the entire SciJava component collection to use
that release.

Unfortunately, the logic was too general: components with no activity
outside a 24-hour window of their last release were erroneously flagged
with None action. We only want to override the action when the vetting
was done manually as a timestamp override.