github.py: handle more than 1000 results
GitHub returns a maximum of 1000 items per search query. But since we
request issues in creation date order, we can use a created:> constraint
to continue the search where the results left off, until we stop hitting
the 1000-item limit. To avoid going off the rails, we set the maximum
number of requests to 100 by default (i.e. 10000 issues total).

We also pass a User-Agent, and wait 5 seconds between each request,
to encourage GitHub not to stop us cold with 403 Forbidden errors.