Update badge loading code to be more robust
I tried quite a lot to keep the previous approach to badge sorting
(i.e. hashing from the image content), but was not able to overcome
CORS failures in all cases when attempting to load and parse the
badge image data asynchronously via JavaScript. So now we use
api.github.com instead to access the last build status that way.

It now requires a GitHub personal access token, unfortunately,
but this way we stay within GitHub's badge access rate limit.